portable/requirements.txt
torch
numpy
pyyaml
opencv-python
albumentations

